Use (u)int64_t instead of (unsigned) long long int in various places

+++ src/pwanfileinfo.cpp
@@ -38,7 +38,7 @@ std::string pwan::fileInfo::path(void)
     return internalpath;
 }
 
-unsigned long long int pwan::fileInfo::size(void)
+uint64_t pwan::fileInfo::size(void)
 {
     return internalsize;
 }
@@ -53,7 +53,7 @@ void pwan::fileInfo::setPath(const std::string& newPath)
     internalpath = newPath;
 }
 
-void pwan::fileInfo::setSize(long long int newSize)
+void pwan::fileInfo::setSize(uint64_t newSize)
 {
     internalsize = newSize;
 }
